Much better 2d Platformer collisions
