+++ src/main.rs
@@ -1,7 +1,8 @@
 #![allow(nonstandard_style)]
 mod Engine;
 mod GrapplerSceneBuilder;
-mod PlayerController;
+mod LunaController;
+mod Grappler;
 
 #[macro_use]
 extern crate glium;
